+++ topicdb/core/commands/association/getassociationgroups.py
@@ -5,21 +5,21 @@ July 13, 2016
 Brett Alistair Kromkamp (brett.kromkamp@gmail.com)
 """
 
-from topicdb.core.models.doublekeydict import DoubleKeyDict
+from topicdb.core.commands.topic.gettopicassociations import GetTopicAssociations
 from topicdb.core.models.associationfield import AssociationField
+from topicdb.core.models.doublekeydict import DoubleKeyDict
 from topicdb.core.topicstoreerror import TopicStoreError
-from topicdb.core.commands.association.getassociations import GetAssociations
 
 
 class GetAssociationGroups:
 
     def __init__(self,
                  database_path='',
-                 map_identifier=None,
+                 topic_map_identifier=None,
                  identifier='',
                  associations=None):
         self.database_path = database_path
-        self.map_identifier = map_identifier
+        self.topic_map_identifier = topic_map_identifier
         self.identifier = identifier
         self.associations = associations
 
@@ -29,13 +29,13 @@ class GetAssociationGroups:
             raise TopicStoreError(
                 "At least one of the 'identifier' or 'associations' parameters is required")
 
-        if self.associations is None and (self.database_path == '' or self.map_identifier is None):
-            raise TopicStoreError("Missing 'database path' or 'map identifier' parameters")
+        if self.associations is None and (self.database_path == '' or self.topic_map_identifier is None):
+            raise TopicStoreError("Missing 'database path' or 'topicmap identifier' parameters")
 
         result = DoubleKeyDict()
         if not self.associations:
-            self.associations = GetAssociations(self.database_path, self.map_identifier,
-                                                self.identifier).execute()
+            self.associations = GetTopicAssociations(self.database_path, self.topic_map_identifier,
+                                                     self.identifier).execute()
 
         for association in self.associations:
             resolved_topic_refs = self._resolve_topic_refs(association)