+++ topicdb/core/commands/association/setassociation.py
@@ -20,11 +20,11 @@ from topicdb.core.topicstoreerror import TopicStoreError
 
 class SetAssociation:
 
-    def __init__(self, database_path, map_identifier,
+    def __init__(self, database_path, topic_map_identifier,
                  association=None,
                  ontology_mode=OntologyMode.strict):
         self.database_path = database_path
-        self.map_identifier = map_identifier
+        self.topic_map_identifier = topic_map_identifier
         self.association = association
         self.ontology_mode = ontology_mode
 
@@ -33,13 +33,13 @@ class SetAssociation:
             raise TopicStoreError("Missing 'association' parameter")
 
         if self.ontology_mode is OntologyMode.strict:
-            instance_of_exists = TopicExists(self.database_path, self.map_identifier,
+            instance_of_exists = TopicExists(self.database_path, self.topic_map_identifier,
                                              self.association.instance_of).execute()
             if not instance_of_exists:
                 raise TopicStoreError(
                     "Ontology mode 'strict' violation: 'instance Of' topic does not exist")
 
-            scope_exists = TopicExists(self.database_path, self.map_identifier,
+            scope_exists = TopicExists(self.database_path, self.topic_map_identifier,
                                        self.association.scope).execute()
             if not scope_exists:
                 raise TopicStoreError(
@@ -49,18 +49,18 @@ class SetAssociation:
 
         try:
             with connection:
-                connection.execute("INSERT INTO topic (topicmap_identifier, identifier, instance_of, scope) VALUES (?, ?, ?, ?)", (self.map_identifier, self.association.identifier, self.association.instance_of, self.association.scope))
+                connection.execute("INSERT INTO topic (topicmap_identifier, identifier, instance_of, scope) VALUES (?, ?, ?, ?)", (self.topic_map_identifier, self.association.identifier, self.association.instance_of, self.association.scope))
                 for base_name in self.association.base_names:
                     connection.execute("INSERT INTO basename (topicmap_identifier, identifier, name, topic_identifier_fk, language) VALUES (?, ?, ?, ?, ?)",
-                                       (self.map_identifier,
+                                       (self.topic_map_identifier,
                                         base_name.identifier,
                                         base_name.name,
                                         self.association.identifier,
                                         base_name.language.name))
                 for member in self.association.members:
-                    connection.execute("INSERT INTO member (topicmap_identifier, identifier, role_spec, association_identifier_fk) VALUES (?, ?, ?, ?)", (self.map_identifier, member.identifier, member.role_spec, self.association.identifier))
+                    connection.execute("INSERT INTO member (topicmap_identifier, identifier, role_spec, association_identifier_fk) VALUES (?, ?, ?, ?)", (self.topic_map_identifier, member.identifier, member.role_spec, self.association.identifier))
                     for topic_ref in member.topic_refs:
-                        connection.execute("INSERT INTO topicref (topicmap_identifier, topic_ref, member_identifier_fk) VALUES (?, ?, ?)", (self.map_identifier, topic_ref, member.identifier))
+                        connection.execute("INSERT INTO topicref (topicmap_identifier, topic_ref, member_identifier_fk) VALUES (?, ?, ?)", (self.topic_map_identifier, topic_ref, member.identifier))
 
             if not self.association.get_attribute_by_name('creation-timestamp'):
                 timestamp = str(datetime.now())
@@ -70,7 +70,7 @@ class SetAssociation:
                                                 scope='*',
                                                 language=Language.eng)
                 self.association.add_attribute(timestamp_attribute)
-            SetAttributes(self.database_path, self.map_identifier,
+            SetAttributes(self.database_path, self.topic_map_identifier,
                           self.association.attributes).execute()
         except sqlite3.Error as error:
             raise TopicStoreError(error)