+++ topicdb/core/commands/attribute/deleteattributes.py
@@ -12,9 +12,9 @@ from topicdb.core.topicstoreerror import TopicStoreError
 
 class DeleteAttributes:
 
-    def __init__(self, database_path, map_identifier, entity_identifier=''):
+    def __init__(self, database_path, topic_map_identifier, entity_identifier=''):
         self.database_path = database_path
-        self.map_identifier = map_identifier
+        self.topic_map_identifier = topic_map_identifier
         self.entity_identifier = entity_identifier
 
     def execute(self):
@@ -25,7 +25,7 @@ class DeleteAttributes:
 
         try:
             with connection:  # https://docs.python.org/3/library/sqlite3.html#using-the-connection-as-a-context-manager
-                connection.execute("DELETE FROM attribute WHERE topicmap_identifier = ? AND parent_identifier_fk = ?", (self.map_identifier, self.entity_identifier))
+                connection.execute("DELETE FROM attribute WHERE topicmap_identifier = ? AND parent_identifier_fk = ?", (self.topic_map_identifier, self.entity_identifier))
         except sqlite3.Error as error:
             raise TopicStoreError(error)
         finally: