+++ topicdb/core/commands/tag/settags.py
@@ -11,11 +11,11 @@ from topicdb.core.commands.tag.settag import SetTag
 
 class SetTags:
 
-    def __init__(self, database_path, map_identifier,
+    def __init__(self, database_path, topic_map_identifier,
                  identifier='',
                  tags=None):
         self.database_path = database_path
-        self.map_identifier = map_identifier
+        self.topic_map_identifier = topic_map_identifier
         self.identifier = identifier
         self.tags = tags
 
@@ -24,4 +24,4 @@ class SetTags:
             raise TopicStoreError("Missing 'tags' or 'identifier' parameter")
 
         for tag in self.tags:
-            SetTag(self.database_path, self.map_identifier, self.identifier, tag).execute()
+            SetTag(self.database_path, self.topic_map_identifier, self.identifier, tag).execute()