+++ topicdb/core/commands/topic/gettopic.py
@@ -18,13 +18,13 @@ from topicdb.core.topicstoreerror import TopicStoreError
 
 class GetTopic:
 
-    def __init__(self, database_path, map_identifier,
+    def __init__(self, database_path, topic_map_identifier,
                  identifier='',
                  resolve_attributes=RetrievalOption.dont_resolve_attributes,
                  resolve_occurrences=RetrievalOption.dont_resolve_occurrences,
                  language=Language.eng):
         self.database_path = database_path
-        self.map_identifier = map_identifier
+        self.topic_map_identifier = topic_map_identifier
         self.identifier = identifier
         self.resolve_attributes = resolve_attributes
         self.resolve_occurrences = resolve_occurrences
@@ -40,13 +40,13 @@ class GetTopic:
 
         cursor = connection.cursor()
         try:
-            cursor.execute("SELECT identifier, instance_of FROM topic WHERE topicmap_identifier = ? AND identifier = ? AND scope IS NULL", (self.map_identifier, self.identifier))
+            cursor.execute("SELECT identifier, instance_of FROM topic WHERE topicmap_identifier = ? AND identifier = ? AND scope IS NULL", (self.topic_map_identifier, self.identifier))
             topic_record = cursor.fetchone()
             if topic_record:
                 result = Topic(topic_record['identifier'], topic_record['instance_of'])
                 result.clear_base_names()
                 cursor.execute("SELECT name, language, identifier FROM basename WHERE topicmap_identifier = ? AND topic_identifier_fk = ?",
-                               (self.map_identifier,
+                               (self.topic_map_identifier,
                                 self.identifier))
                 base_name_records = cursor.fetchall()
                 if base_name_records:
@@ -56,10 +56,10 @@ class GetTopic:
                                      Language[base_name_record['language']],
                                      base_name_record['identifier']))
                 if self.resolve_attributes is RetrievalOption.resolve_attributes:
-                    result.add_attributes(GetAttributes(self.database_path, self.map_identifier,
+                    result.add_attributes(GetAttributes(self.database_path, self.topic_map_identifier,
                                                         self.identifier).execute())
                 if self.resolve_occurrences is RetrievalOption.resolve_occurrences:
-                    result.add_occurrences(GetOccurrences(self.database_path, self.map_identifier,
+                    result.add_occurrences(GetOccurrences(self.database_path, self.topic_map_identifier,
                                                           self.identifier).execute())
         except sqlite3.Error as error:
             raise TopicStoreError(error)